+++ packages/client/ui/react-ui/src/components/index.ts
@@ -3,5 +3,6 @@ export * from "./CrossmintNFTDetail";
 
 export * from "./embed";
 export * from "./hosted";
+export * from "./embed/v3";
 
 export { EmbeddedAuthForm } from "./auth/EmbeddedAuthForm";