+++ packages/common/auth/src/types/authmaterial.ts
@@ -5,14 +5,12 @@ export type AuthMaterialBasic = {
     refreshToken: string;
 };
 
-interface RefreshToken {
-    secret: string;
-    expiresAt: string;
-}
-
 export type AuthMaterial = {
     jwt: string;
     refreshToken: RefreshToken;
+};
+
+export type AuthMaterialWithUser = AuthMaterial & {
     user: SDKExternalUser;
 };
 
@@ -24,7 +22,13 @@ export type AuthMaterialResponse = {
 
 export type AuthSession = {
     jwt: string;
+    refreshToken: RefreshToken;
     userId: string;
 };
+
+interface RefreshToken {
+    secret: string;
+    expiresAt: string;
+}
 
 export type OAuthProvider = "google";